+++ ocdskit/cli/commands/indent.py
@@ -37,7 +37,7 @@ class Command(BaseCommand):
                 data = json.load(f)
 
             with open(path, 'w') as f:
-                json_dump(data, f, indent=self.args.indent)
+                json_dump(data, f, indent=self.args.indent, ensure_ascii=self.args.ascii)
                 f.write('\n')
         except json.decoder.JSONDecodeError as e:
             logger.error('{} is not valid JSON. (json.decoder.JSONDecodeError: {})'.format(path, e))